♻️ refactor: 답변 내용 부분을 FeedCardAnswer로 올림 / FeedCardAnswerContent > FeedCardAnswerInput
답변 상태에 따라 보여주는 답변 내용을 달리 하는 부분을 FeedCard로 올림
